+++ TotalImage.IO/FileSystems/ISO/Iso9660FileSystem.cs
@@ -36,7 +36,7 @@ namespace TotalImage.FileSystems.ISO
         public override long TotalSize { get; }
 
         /// <inheritdoc />
-        public override uint AllocationUnitSize => PrimaryVolumeDescriptor.LogicalBlockSize;
+        public override long AllocationUnitSize => PrimaryVolumeDescriptor.LogicalBlockSize;
 
         /// <inheritdoc />
         public override bool SupportsSubdirectories => true;
